FDK-326 Oppgradere Spring Boot til 1.5.1  - Moved cucumber model objects to model  - Replaced local-properties with application.yml  - Development profile is now develop for all modules
