+++ harvester/src/main/java/no/dcat/harvester/settings/ApplicationSettings.java
@@ -1,13 +1,8 @@
 package no.dcat.harvester.settings;
 
-import org.springframework.beans.factory.config.YamlPropertiesFactoryBean;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
-import org.springframework.core.io.ClassPathResource;
 
 @Configuration
 @EnableConfigurationProperties
@@ -46,19 +41,6 @@ public class ApplicationSettings {
 
 	public void setCrawlerThreadPoolSize(int crawlerThreadPoolSize) {
 		this.crawlerThreadPoolSize = crawlerThreadPoolSize;
-	}
-
-	//Det må lages en egen PropertySourcePlaceholderConfigurer siden @PropertySource fortsatt ikke støtter yaml format.
-	@Bean
-	@Profile("development") //Skal kun brukes når spring_active_profiles inneholder default
-	public static PropertySourcesPlaceholderConfigurer properties() {
-		PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer = new PropertySourcesPlaceholderConfigurer();
-		YamlPropertiesFactoryBean yaml = new YamlPropertiesFactoryBean();
-
-		//Path til propertiesfiler som skal brukes for JUnit og kjøring på lokal maskin
-		yaml.setResources(new ClassPathResource("properties/local-properties.yml"));
-		propertySourcesPlaceholderConfigurer.setProperties(yaml.getObject());
-		return propertySourcesPlaceholderConfigurer;
 	}
 
 }
