+++ harvester/src/main/resources/application.yml
@@ -1,28 +1,21 @@
-fuseki:
-  dcatServiceUri: http://localhost:3030/fuseki/dcat
-  adminServiceUri: http://localhost:3030/fuseki/admin
-application:
-  crawlerThreadPoolSize: 2
-  elasticSearchHost: localhost
-  elasticSearchPort: 9300
-  elasticSearchCluster: elasticsearch
----
 spring:
     profiles: docker
 fuseki:
   dcatServiceUri: http://fuseki:3030/fuseki/dcat
   adminServiceUri: http://fuseki:3030/fuseki/admin
 application:
+  crawlerThreadPoolSize: 2
   elasticsearchHost: es
   elasticsearchPort: 9300
   elasticsearchCluster: elasticsearch
 ---
 spring:
-    profiles: development
+    profiles: develop
 fuseki:
   dcatServiceUri: http://localhost:3030/fuseki/dcat
   adminServiceUri: http://localhost:3030/fuseki/admin
 application:
+  crawlerThreadPoolSize: 2
   elasticsearchHost: localhost
   elasticsearchPort: 9300
   elasticsearchCluster: elasticsearch
