+++ harvester/src/main/resources/banner.txt
@@ -16,15 +16,15 @@
                ........................................................................     /
                 .........................................................................    /
                 ...........................................................................  /
-                          .............................................%%%%%%%%............. /
-                            .........................................%###((((###%.............*
-                      ..............................................%#((//////((#%................   .....
-                    .....*%%(......................................(##(/*....*/(##(...................... ,
-                   ....(#(//(#%....................................*%#(//*..*//(#%*....    ............... ,
-                   ....#(/..*(#*....................................#%#(((((((##%#.....      ............,,,,.,
-                   ....##(//(#%....                             ......%%%####%%%......          ........,,,,,, ,,
-                    .....(%%#.....                               .........,,.........               ... ,,,,,,.,,
-                      ..........                                   ................                     .,,,,,,,,,,,
+                          ..................................../  ......%%%%%%%%............. /
+                            ................................./  .....%###((((###%.............*
+                      .......................................| .....%#((//////((#%................   ....\
+                    .....*%%(................................| ....(##(/*....*/(##(.......................|
+                   ....(#(//(#%..............................\ ....*%#(//*..*//(#%*....    ................\
+                   ....#(/..*(#*..............................| ....#%#(((((((##%#.....      ............,,,,.\
+                   ....##(//(#%....                             ......%%%####%%%......          ........,,,,,, ,\
+                    .....(%%#.....                               .........,,.........               ... ,,,,,,.,,\
+                      ..........                                   ................                     .,,,,,,,,,,\
                 __  __                           __
                / / / /___ _______   _____  _____/ /____  _____
               / /_/ / __ `/ ___/ | / / _ \/ ___/ __/ _ \/ ___/
