--- portal/query/src/test-local/resources/properties/local-properties.yml
@@ -1,24 +1,0 @@
-logging:
-  level.: ERROR
-  level.no.brreg: DEBUG
-  level.org.springframework: ERROR
----
-spring:
-    profiles: docker
-application:
-  elasticsearchHost: elasticsearch
-  elasticsearchPort: 9300
-  clusterName: elasticsearch
----
-#Utv-miljø lokalt på egen PC
-spring:
-  profiles: development
-application:
-  elasticsearchHost: localhost
-  elasticsearchPort: 9300
-  clusterName: elasticsearch
-server:
-  port: 8083
-management:
-  port: 8993
-
