--- portal/webapp/src/test-local/resources/properties/local-properties.yml
@@ -1,24 +1,0 @@
-spring:
-  profiles: docker
-application:
-  queryService: http://localhost:8081
-  queryServiceExternal: http://localhost:8081
-  fusekiService: http://localhost:3030/fuseki
----
-spring:
-  profiles: test
-application:
-  queryService: http://localhost:8071
-  queryServiceExternal: http://localhost:8071
-  fusekiService: http://localhost:3030/fuseki
----
-spring:
-  profiles: develop
-application:
-  queryService: http://localhost:8083
-  queryServiceExternal: http://localhost:8083
-  fusekiService: http://localhost:3030/fuseki
-server:
-  port: 8080
-management:
-  port: 9090
